+++ apps/router-e2e/__e2e__/server-loader/app/index.tsx
@@ -1,0 +1,24 @@
+import { Text, View } from "react-native";
+import { useLoader } from "expo-router";
+
+export async function loader() {
+  return Promise.resolve({
+    foo: 'bar',
+  });
+}
+
+export default function Index() {
+  const data = useLoader(loader);
+
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <Text testID="loader-result">{JSON.stringify(data)}</Text>
+    </View>
+  );
+}