+++ packages/@expo/cli/CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add `EXPO_UNSTABLE_LIVE_BINDINGS` to allow developer to disable live binding in `experimentalImportSupport`. ([#38135](https://github.com/expo/expo/pull/38135) by [@krystofwoldrich](https://github.com/krystofwoldrich))
 - Add `location.origin`, Expo SDK version and Hermes version to sitemap UI ([#38201](https://github.com/expo/expo/pull/38201) by [@hassankhan](https://github.com/hassankhan))
 - Allow running server middleware with `+middleware.ts` ([#38330](https://github.com/expo/expo/pull/38330) by [@hassankhan](https://github.com/hassankhan))
+- [web] Add server data loader support in static export mode ([#38524](https://github.com/expo/expo/pull/38524) by [@hassankhan](https://github.com/hassankhan))
 
 ### 🐛 Bug fixes