+++ packages/expo-router/CHANGELOG.md
@@ -22,6 +22,7 @@
 - Native bottom tabs ([#37912](https://github.com/expo/expo/pull/37912) by [@Ubax](https://github.com/Ubax))
 - Allow running server middleware with `+middleware.ts` ([#38330](https://github.com/expo/expo/pull/38330) by [@hassankhan](https://github.com/hassankhan))
 - NativeTabs dynamic options ([#38581](https://github.com/expo/expo/pull/38581) by [@Ubax](https://github.com/Ubax))
+- [web] Add server data loader support in static export mode ([#38524](https://github.com/expo/expo/pull/38524) by [@hassankhan](https://github.com/hassankhan))
 
 ### 🐛 Bug fixes